Remove unused imports, move get_server_and_client function, remove unused exception class, comment cleanups, misc formatting changes
