+++ mangum/types.py
@@ -9,7 +9,6 @@ except ImportError:
     # Otherwise, Protocol has incompatible type "typing_extensions._SpecialForm"
 
 Message = typing.Dict[str, typing.Any]
-
 Scope = typing.Dict[str, typing.Any]
 Receive = typing.Callable[[], typing.Awaitable[Message]]
 Send = typing.Callable[[Message], typing.Awaitable[None]]
