+++ mangum/utils.py
@@ -11,29 +11,6 @@ def make_response(content: str, status_code: int = 500) -> dict:
     }
 
 
-def get_server_and_client(event: dict) -> typing.Tuple:  # pragma: no cover
-    """
-    Parse the server and client for the scope definition, if possible.
-    """
-    client_addr = event["requestContext"].get("identity", {}).get("sourceIp", None)
-    client = (client_addr, 0)
-
-    server_addr = event["headers"].get("Host", None)
-
-    if server_addr is not None:
-        if ":" not in server_addr:
-            server_port = 80
-        else:
-            server_addr, server_port = server_addr.split(":")
-            server_port = int(server_port)
-
-        server = (server_addr, server_port)  # type: typing.Any
-    else:
-        server = None
-
-    return server, client
-
-
 def get_logger(log_level: str) -> logging.Logger:
     level = {
         "critical": logging.CRITICAL,
